extract logic to concern
